Tweaked KeyAction GSON serialization:
Created InterfacetoJson Json adapter
Created JsonAdapterRegistrar to register adapters with gson.
Created a Convertable marker interface to mark interfaces for convering
Tweaked KeyPreferenceIO to work with new adapter and registrar
Tweaked DefaultFileCreator to work with new adapter
Regenerated keybind.config
Cleaned up various classes
Generated new Javadoc
Various bug fixes

+++ src/SwissCheese/engine/camera/Camera.java
@@ -29,11 +29,11 @@ import SwissCheese.map.Map;
  * @version v0.2
  */
 public class Camera {
-	private final int FOV;
+	private final float FOV;
 	private Mover mover;
 	private View view;
 
-	public Camera(int width, int height, Map map, Point2D start, int FOV) {
+	public Camera(int width, int height, Map map, Point2D start, float FOV) {
 		this.FOV = FOV; //0 -> -1
 		view = new View((int)start.getX(), (int)start.getX(), 1, 0, 0, FOV);
 		mover = new Mover(view, map);
@@ -48,7 +48,7 @@ public class Camera {
 		return mover;
 	}
 
-	public synchronized final int getFOV() {
+	public synchronized final float getFOV() {
 		return FOV;
 	}
